Change KotlinTypeResolver to function
